KAN-27-fix:swagger 비인증 url 추가
fix:swagger 비인증 url 추가

swagger 비인증 url 추가

+++ announcement/src/main/java/org/pda/announcement/stockprice/domain/Stockpriceday.java
@@ -46,7 +46,7 @@ public class Stockpriceday {
     private Integer closePrice;
 
     @Column(name = "volume")
-    private Integer volume;
+    private Long volume;
 
     @Column(name = "change_rate")
     private Float changeRate;
